+++ lib/src/models/translation_model.dart
@@ -1,0 +1,6 @@
+class TranslationModel {
+  final String translatedText;
+  final String detectedSourceLanguage;
+
+  TranslationModel({required this.translatedText, required this.detectedSourceLanguage});
+}