+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.0.2] - (2021-Sep-13).
+
+* Breaking changes. Now result of translation returns with TranslationModel with two properties
+  - translatedText
+  - detectedSourceLanguage
+  This is helps to detect cases when language we translate To is same as the source.
+
 ## [0.0.1+1] - (2021-Aug-8).
 
 * Fixed readme & description.